+++ tests/tempo/backends/base_backends_test.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-Tests for the oqupy.backends.base_backends module.
+Tests for the oqupy.tempo.backends.base_backends module.
 """
 import pytest
 
@@ -20,10 +20,22 @@ import numpy as np
 import tensornetwork as tn
 
 import oqupy as tempo
-from oqupy.backends.backend_factory import \
-    get_tempo_backend, get_pt_tempo_backend
+import oqupy.tempo.backends.base_backends as bb
 pass
 
-def test_backend_factory_default():
-    tempo_back = get_tempo_backend()
-    pt_tempo_back = get_pt_tempo_backend()
+def test_base_tempo_backend():
+    tempo_back = bb.BaseTempoBackend(
+                    initial_state=None,
+                    influence=None,
+                    unitary_transform=None,
+                    propagators=None,
+                    sum_north=None,
+                    sum_west=None,
+                    dkmax=None,
+                    epsrel=None,
+                    config=None)
+    with pytest.raises(NotImplementedError):
+        tempo_back.initialize()
+    with pytest.raises(NotImplementedError):
+        tempo_back.compute_step()
+
